+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* Updated `ViewerSceneObject` with hash ability in order to catch the update from `settings`.
 
 ### Removed
 
@@ -33,8 +34,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed `Tag` inconsistent height issue.
 * Dynamically adjust camera pan delta based on distacne.
 
-* Updated alignright feature to `TextEdit` 
-
 ### Removed
 
 
@@ -54,7 +53,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Updated callback to `SceneTree`.
 * Updated `ObjectSetting` and `CameraSetting` to support setting from config.
 * Updated `Slider` to be able change value with `TextEdit`
-* Updated `ViewerSceneObject` with hash ability to check if `kwarg` is equal 
 
 ### Removed